+++ index.js
@@ -4,7 +4,7 @@
 Object.assign(navigator || {}, { product: 'ReactNative' })
 
 import '@walletconnect/react-native-compat'
-import './shim'
+import './polyfills/shim'
 import 'react-native-get-random-values'
 import '@ethersproject/shims'
 import 'react-native-crypto'
@@ -12,7 +12,7 @@ import 'text-encoding-polyfill'
 import 'intl'
 import 'intl/locale-data/jsonp/en'
 import 'react-native-gesture-handler'
-import './rn-polyfill-deprecated-view-prop-types'
+import './polyfills/rn-deprecated-view-props-types-polyfill'
 
 import { AppRegistry, LogBox } from 'react-native'
